Rename Pathfollower.js to PathFollower.js
The file /plugins/pathfollower.js contains the following import statement:
import PathFollower from './behaviors/pathfollower/PathFollower.js';
This references the changed file with a capital F, however the current version of this file is named Pathfollower.js with a lowercase f. This causes bundlers to fail on case-sensitive platforms such as linux. This change fixes the issue.